Remove ImageDriver (and related plugin) code
In 3.11 we deprecated ImageDriver plugins. The only ImageDriver
plugin distributed with SingularityCE was an example used to allow
unprivileged overlay on Ubuntu's custom kernel. Now that unprivileged
overlay is a mainline kernel feature, support has been implemented
for unprivileged overlay without a plugin.

SingularityCE did not use the broader ImageDriver construct outside of
plugin handling. The code was also mostly un-tested. Removing it
simplifies various functions in the native runtime engine.

If we move to fully support FUSE for squashfs/extfs/overlay in the
native runtime, then our wish is to do so using implementation code
that can be heavily shared with the OCI runtime, e.g. the new
OverlaySet etc. This approach does not fit well with the ImageDriver
concept, as it was tightly coupled to the native runtime engine.

Closes #912