Enter namespaces earlier in --oci mode.
Before this PR, if we are a non-root user we enter a root-mapped user
namespace before calling `singularity oci run` at the end of the
`--oci` launcher flow. If we are the root user, we do not enter any
namespace. This means that bundle preparation, performed by the
launcher, takes place outside of any namespaces.

As a non-root user, we want to perform bundle preparation inside a
root-mapped user namespace. This is in order to benefit from
fully-unprivileged FUSE operations when we start handling image
mounts, plus advantages around cleanup on namespace exit. As a root
user, we want to be inside a mount namespace to avoid polluting the
host mount namespace with our container-specific mounts.

In this PR:

* As a non-root user, a root-mapped namespace is entered early, before
  the oci launcher code executes.
* As a root user, a mount namespace is entered early, before the oci
  launcher code executes.

We use the `_CONTAINERS_ROOTLESS_xID` environment variables for
tracking our UID/GID outside of the root-mapped namespace, since they
must be set anyway for `containers/image` to operate properly inside
the namespace. We are using `containers/image` for our OCI image handling.

Fixes #1773