fix: disable IDMapping in inter-stage copy for proot builds (release-3.11)
In builds, we use a wrapped `CopyWithTar` to handle `%file` copies
between stages. The wrapper includes logic to perform ID mappings to
ensure correct, squashed, ownership of files.

Within a `proot` build, a copy between stages has been found to
hang. The ID mapping is the cause of the hang. Disable it for the
particular circumstance of proot build between-stage copy.

Fixes #1643