oci: switch direct run/shell/exec from URI to oci-sif flow
When a direct URI is provided for `run/shell/exec` in `--oci` mode,
perform a pull through the cache to obtain an OCI-SIF image, and
execute this.

`--oci` mode is no longer limited to OCI URIs or `oci-sif` as
valid sources. If an `oci-sif` can be retrieved from the
library/shub/oras/http/https protocols it can be run.

At this point, http/https is trivial to e2e-test so can be supported.

The barriers for the other protocols are primarily around pushing an
image, rather than pulling. When push of oci-sif is supported and
tested in follow-up PRs, then we can list the protocols as supported.

Closes #1860