oci: re-exec as early as possible
When we need to enter a mount or user namespace to begin oci-mode
operation, do so as early as possible.

The CLI code has a lot of logic in it, much of which may run twice if we
re-exec in `launchContainer`. This code may also cause side effects.

Re-execing much earlier, at the root command's `PersistentPreRun`,
avoids these issues.