pkg/image: Add minimal oci-sif type
Add a minimal oci-sif image type to pkg/image, that does not implement
any functionality other than oci-sif detection.

Ensure oci-sif detected as oci-sif, and are not detected as the
existing sif type.

Modify the native launcher to return a more friendly error if a user
attempts to run an oci-sif with it.

Prep work for #1861