oci: Execute non-oci SIF in --oci mode
Implement basic execution of non-oci SIF images in `--oci` mode.

In the OCI launcher we now expect the process argument to `Exec` to be
an action script, and strip it out if we are *not* running a native SIF.

Handling of `library://`, `oras://` is already present via pull
through the cache.

Closes #1899