library: push/pull OCI-SIF images
When `singularity push` is provided an OCI-SIF image file and a
`library://` URI, push as a single layer squashfs OCI image to the OCI
registry that backs the library service. Existing push behaviour for
native SIF through the library API remains.

When `singularity pull --oci` is provided a `library://` URI with a
path that points to a single layer squashfs OCI image in the backing
OCI registry, pull to a local OCI-SIF image.

When `singularity pull` is provided a `library://` URI and no `--oci`
flag, try a native SIF library API pull first. If that fails and the
error indicates the image is an OCI image then try an OCI-SIF pull.

The push/pull process involves discovering the URI of the backing OCI
registry, retrieving an authentication token, and translating
`library://` URIs to OCI registry URIs.

Because a `pull` of an standard OCI image to a native SIF is
implemented as a build, and the build code imports from
`pkg/client/*`, we have to do some gymnastics to avoid introducing
import loops:

* Code dealing with oci image operations used by both builds and the
  oci client package has been moved to `internal/pkg/ociimage`.

* Low-level push/pull between OCI-SIF and OCI registries has been
  placed in `internal/pkg/client/ocisif` rather than `oci`.

Closes #1916