pull: Pull previously pushed OCI-SIF from OCI registry (docker://)
If `singularity pull --oci` is used against an OCI registry
`docker://` reference, and the reference is a single-layer squashfs
image, pull it into an OCI-SIF without any mutation.

This is the mirror-image operation of #1917, allowing OCI-SIF images
pushed with the code in #1917 to be pulled.

Note that the pull e2e tests have been re-arranged / re-named. This is
to keep all the `library://` `oras://` `docker://` tests next to
eachother.... and in an attempt to make it more obvious what is
testing what.

Closes #1915