remote: Make registryUri available for library
Retrieve the `registryUri` value from the library service config for
SCS/Enterprise remotes.

The concept of a service in the remote code is rather overloaded and
inflexible. Add an internal use map that we can pull the `registryUri`
into and out of. Hide it behind a single `RegistryURI` public fn.

Not the tidiest, but it works pending major overhaul of the remote /
endpoint code.