chore: remove pkg/ociruntime
As we are bumping to a new major version, we can remove pkg/ociruntime
at this point.

We don't have our own implementation of an OCI runtime since 3.10.0
and the code here is only used in e2e-tests.

It's better to use the upstream oci runtime-spec structs / constants
instead, in those tests.

Part of #80