oci: Implement --allow-setuid, match nosuid mount to native mode
In native mode, things are nosuid mounted by default, and
`--allow-setuid` is available to the root user to relax this.

In OCI mode, the `nosuid` option on mounts is not security critical as
when run by an unpriv user, we always perform the mounts within a
root-mapped user namespace.

Regardless, match the behavior with native mode by default, and allow
any user to specify `--allow-setuid` to remove the `nosuid` from
appropriate mounts.

Includes some tidy up of overlay item option handling, and adds tests
to check mount options on overlay items.

Closes #1469