oci: enable signal propagation
When we are using os.exec in the OCI-mode execution flow, we must
handle any signals we receive, and proxy them to the container.

For non-root users, we are using the fakeroot starter (which already
proxies signals) to re-exec singularity in a userns. Proxy signals to
the container via `runc/crun kill`.

For root users, we are using os.Exec to re-exec singularity in a
userns, so we need to add singal proxying across this boundary in
`rootless.RunInMountNS`.

Also, we need to propagate signals to exit codes correctly.

Fixes #1779