no-compat: bind full /dev
When `--oci` mode is run with `--no-compat`, follow native mode
default behaviour of binding the entire `/dev` from the host into the
container by default.

If `mount dev = minimal` in `singularity.conf` then we don't mount the
entire host `/dev`, but use the minimum for an OCI container.

If `mount dev = no` in `singularity.conf` or `--no-mount dev` supplied
then report an error as we require dev. Error now, instead of warn, so
it's clear we can't acheive the native mode behavior.

If `mount devpts = no` in `singularity.conf` or `--no-mount devpts`
supplied then report an error as we require devpts. Error now, instead
of warn, so it's clear we can't acheive the native mode behavior.

Fixes #1981