oci: allow override of native SIF container env vars
When entering a native SIF container, we apply the container-defined
environment by sourcing scripts in `/.singularity.d/env` from the
relevant action script.

Any environment variable set in these scripts will replace any value
set from the OCI runtime config.

To allow `SINGULARITYENV_` / `--env` / `--env-file` to override
container-defined env vars, we have to inject them via a script that
exports them in `/singularity.d/env`.

We use the script filename `98-singularityenv.sh` as it must:

* Run *after* `95-apps.sh` which sets SCIF App specific
  variables (which we should be able to override).
* Run *before* `99-base.sh` which mutates `LD_LIBRARY_PATH`.

In addition, we have to consider that:

* In OCI mode, by default we emulate native mode `--compat`. This
  implies `--no-eval`, where environment variables injected should not
  be subjected to shell evaluation.
* We have a `--no-compat` flag, that needs to emulate native mode
  defaults. In this case, environment variables injected should be
  subjected to shell evaluation.

In general... this is painfully complex, relies on knowledge of where
/ when various scripts are executed, and may not have the best code
structure. However, it works and, the tests enforce the correct
behavior. Let's tidy up further later if that's okay? :-)

Fixes #1990