fix: eval --env-file w.r.t. host env in --oci mode
In native mode, the contents of an --env-file is a shell script
evaluated with respect to the host environment. In --oci mode, to
date, it has been evaluated with respect to a clean environment.

Refactor the basic environment variable handling out of the OCI
launcher and into the common env package.

Ensure we go through the same code path to evaluate --env-file in
native and OCI modes. Use the current host environment for evaulation
in OCI mode.

Fixes #2005