oci: host environment handling for --no-compat
In `-oci` mode, make `--no-compat` handle host environment in the same
manner as the native runtime.

Most environment variables on the host are passed into the
container, as long as they haven't been overridden by the container
image or the user.

For OCI-SIF images, this is handled when computing the runtime config.

For native SIF images, this is handled by injecting the variables
through a shell script that runs in `/.singularity.d/env` processing.

Note that we now always set `TERM` and proxy environment variables in
the container using these mechanisms, regardless of `--no-compat`. The
proxy env vars are important for frictionless use of Singularity in
isolated environments.

Fixes #1979