conf: Add 'oci mode' directive to singularity.conf
Add an `oci mode` directive to singularity.conf that makes Singularity
default to OCI-mode.

Add a `--no-oci` flag to cancel the effect of the `oci mode`
directive, in order to run / pull native mode / non-OCI SIF container
images.

Refactor `--oci` to be a common flag used by multiple commands.

Tidy up the order / comments on common flags to make it easier to see
what they apply to.

Add some additional guarding of `build` and `instance` commands which
do not have an implementation in OCI mode. Fail with an error if
`--oci` mode is in use on these commands to avoid user believing that
they are running in OCI-mode.

Fixes #1865