oci/cache: Support pull by --platform to OCI-SIF
Adds a `--platform` flag to `singularity pull` that is effective when
pulling from an OCI source to OCI SIF. Note that this includes pulling
from the library in `--oci` mode, as this is an OCI registry operation.

The `--platform` flag can be used to request that the image pulled
satisfies a specific `os/architecture/variant` specification.

I have not overloaded the `--arch` flag used for `library://` pulls,
as `--platform` gives compatibility with OCI runtimes, and the library
`--arch` only supports the single architecture component of a
platform.

A majority of the work involved in this PR is in making our OCI
caching architecture aware.

When the OCI source is not a docker:// remote registry, we always:

 * Retrieve the image manifest or image index at the remote.
 * If we have an image manifest, use its digest as our caching key,
   directly.
 * If we have an image index, use the digest (if any) of a listed
   manifest that satisfies our platform requirements.
 * Add the image index or manifest as a blob in our cache.

When the OCI source is a docker:// registry, handling is more
complex. We need to avoid retrieving the manifest each time as these
`GET` requests count against Docker Hub API limits.

 * Perform a HEAD against the registry. This returns the digest of the
   image manifest or image index.
 * If we have not already cached the matching image manifest or image
   index, proceed as for other sources, as above.
 * If we have already cached the matching image manifest or image
   index, retrieve it from the cache.
 * If it's an image manifest, use its difest as our caching key,
   directly.
 * If we have an image index, use the digest (if any) of a listed
    manifest that satisfies our platform requirements.

The code here is fairly messy, as we have a mix of containers/image
and go-containerregistry.

For 4.1 we have the intent to migrate from containers/image ->
go-containeregistry completely. Consequently I've used a mix that is
the easiest to work with. containers/image platform matching is an
internal package - which makes ggcr useful here.

Note also the addition of code from `containerd/containerd` to
correctly compute the system platform. This is added to our repo under
a compatible license to avoid bringing back a large direct import that
generates a lot of dependabot noise.

Fixes: #1934