fix: prevent scope error with docker io registry login
Use empty scopes for the registry ping after `singularity registry
login`, as a "" scope is resolving to a registry level scope that
gives an error when logging in to Docker Hub.

May result in the credentials not being properly validated, i.e. login
with bad password succeeding. However, this is the best we can do at
the moment with gccr - after extensive investigation by tri-adam and
wobito.