e2e: gate some oci-sif pull tests by arch
At present, we only have single-architecture OCI repositories in the
library, for oci-sif images.

Gate the `library_oci-sif_fallback` and `library_oci-sif_direct` tests
for `amd64`, as source in the library is `amd64` only.

We do not need to gate `library_oci-sif_platform` and
`library_oci-sif_arch` as they are specifying `ppc64le` and using a
`ppc64le` source. The fact they fail on ARM64 is due to a different
issue - #2049

Fixes #2048