oci: handle variant correctly for custom arch
In SysCtxToPlatform, if the `ArchitectureChoice` requested does not
match the host architecture then the variant must not default to the
host variant.

This error is caught in existing unit and e2e tests on ARM64.

Prior to this PR, requesting `--platform linux/ppc64le` leads to a
`linux/ppc64le/v8` platform as the host variant is incorrectly carried
over.

After this PR, requesting `--platform linux/ppc64le` leads to a
`linux/ppc64le` platform. The host variant is dropped because the
architecture doesn't match the host architecture.

Fixes #2049