fix: e2e: handle max cap in OCI capabilities tests
When we are checking for capabilities in the test, we need to
accommodate systems that don't have CAP_AUDIT_READ, e.g. EL7.

CAP_AUDIT_READ is in the default set for an OCI container, but is not
present on EL7. Perform a bitwise and with the system full capset for
the values we are checking.

Fixes #2072