fix: rpm: use semver version-release for mconfig -V
Use `@PACKAGE_VERSION@-@PACKAGE_RELEASE@` as the version specifier in
the call to `mconfig -V` in our rpm spec.

The RPM spec's own `%{release}`, which will still be used for the name
of the RPM package itself, includes `%{?dist}` which may not be semver
compliant.

Additionally, using `@PACKAGE_VERSION@` instead of `%{version}`
preseves semver style `-` for pre-releases instead of rpm style `~`
for pre-releases in the `singularity --version / version`.

We are aiming for semver compliance - our reported version number
should be.

Fixes #2073