fix: Don't mount /var/tmp pointing to /tmp in container
It is not uncommon for `/var/tmp` in a container to be a symlink onto
`/tmp`. In this case, singularity will currently:

* Mount host `/tmp` onto the container `/tmp`
* Mount host `/var/tmp` onto the container `/tmp` through the symlink.

Check whether `/var/tmp` resolves to the same place as `/tmp` in the
container, and do not mount it if this is the case.

We have to move `addTmpMount` to run under a `RunBeforeTag` call
because we need to have the container rootfs in place to to inspect
for symlinks. Because the sources of the mounts may be a
tmpfs (--contain), or inside a --workdir, we still have to handle
creation of the source dirs earlier.

Fixes #1950
Fixes #1331