chore: re-enable dependabot on vault/release-3.11
The 3.11 version of SingularityCE is packaged, by ourselves, in EPEL.

Our intention is to minimally maintain it until a 3.11 version of
SingularityCE falls out of current supported Fedora releases in
approx. 1 year.

Re-enable dependabot to make it easier to track security related
dependency updates.

Note that this branch remains under the `vault/` prefix, because:

* We will not be backporting bugfixes, outside of anything required
  through Fedora/EPEL.
* We will not tagging or performing any further releases from GitHub.
* We will only be modifying code to deal with security issues,
  propagating the changes to EPEL packages.