oci: pull and run multi-layer OCI-SIF images
When an OCI-SIF is created explicitly with `pull`, or implicitly with
`run/shell/exec/instance start`, permit the image layers to be kept
intact.

The new `--keep-layers` flag for these commands skips squashing images
to a single layer when an OCI-SIF is created from an OCI source.

To allow execution of the resulting multi-layer oci-sif files:

* Mount each layer from an oci-sif using squashfuse, onto a `layer/n`
directory in the bundle.
* Mount a read-only overlay using all layers onto the `rootfs`
directory in the bundle.
* Keep track of the mounts so they can be unmounted on `.Delete()`.

Fixes #2266
Fixes #2267