oci-sif: add `--keep-layers` to preserve layers in OCI-SIF creation
When an OCI-SIF is created explicitly with `pull`, or implicitly with
`run/shell/exec/instance start`, permit the image layers to be kept
intact.

The new `--keep-layers` flag for these commands skips squashing images
to a single layer when an OCI-SIF is created from an OCI source.

This is the initial step for multi-layered OCI-SIF creation. It is not
possible to execute the images. Testing will be added as we proceed to
refactor a lot of image handling code to allow execution of these
multi-layer images.

```
$ singularity pull --oci --keep-layers docker://golang
FATAL:   Image file already exists: "golang_latest.oci.sif" - will not overwrite

$ singularity sif list golang_latest.oci.sif
------------------------------------------------------------------------------
ID   |GROUP   |LINK    |SIF POSITION (start-end)  |TYPE
------------------------------------------------------------------------------
1    |1       |NONE    |32176-47541680            |OCI.Blob
2    |1       |NONE    |47541680-66674096         |OCI.Blob
3    |1       |NONE    |66674096-124489136        |OCI.Blob
4    |1       |NONE    |124489136-214961584       |OCI.Blob
5    |1       |NONE    |214961584-281157040       |OCI.Blob
6    |1       |NONE    |281157040-281161136       |OCI.Blob
7    |1       |NONE    |281161136-281162124       |OCI.Blob
8    |1       |NONE    |281162124-281163339       |OCI.Blob
9    |1       |NONE    |281163339-281163580       |OCI.RootIndex
```

Fixes #2266