chore: replace UnavailableError with wrapping for i/p/client/ocisif
In pkg/ocibundle/ocisif an UnavailableError type was provided to
hold an underlying error, effectively performing error wrapping. The
error was used:

* In pkg/ocibundle/ocisif to wrap errors from a failed squashfs mount.
* In internal/pkg/client to wrap errors from a failed tar -> squashfs
conversion.

The underlying intent was to use this error type to implement fallback
in the actions flows from OCI-SIF create/execution to a directory
bundle.

Sharing a generic error across the packages isn't ideal, and we should
be able to use Go's native error wrapping.

However, we shouldn't change `pkg/ocibundle/ocisif` as this modifies a
public API.

This PR replaces the UnavailableError type in the client case with a
specific error, and the use of error wrapping.

It was also noted that in actions.go error messages / warning were
very long and confusing to read. They have been split to multiple
errors / warnings.