fix: flatten overlay mounts in OCI launcher
When a user specifies a read-only `--overlay` but no writable
'--overlay', successful container startup requires that we add an
ephemeral writable-tmpfs that runc / crun can write into.

Prior to this PR, '--overlay' overlays were handled and then the
ephemeral writable-tmpfs was added in a separate overlayfs mount.

We can avoid 2 stacked overlayfs mounts by adding the ephemeral
writable-tmpfs directly within 'WrapWithOverlays'.

This flattening is required so we don't exceed the 2 overlayfs
stacking limit with multi-layer OCI-SIF images, which require an
overlayfs mount to assemble the initial rootfs.