refactor: use ggcr for ociimage.ImageDigest
Re-implement the OCI image digest flows using ggcr instead of
containers/image/v5.

When obtaining the ImageDigest for an image in a remote registry we
need to try to use a `HEAD` request that doesn't count against Docker
Hub API limits.

Unfortunately, the `Docker-Content-Digest` returned by `HEAD` is for
the image index when a repository is multi-arch. We have to locally
cache the manifest or image index so that we can always resolve the
digest we get back from `HEAD` requests to a true image digest.

Fixes #2383