build: split buildkitd code into separate singularity-buildkitd
Prior to this PR, the `singularity` binary included a fairly complete
`buildkitd` implementation.

This PR splits out the buildkit daemon code so that it is compiled
into a separate `singularity-buildkitd` executable, executed as a
child process when required.

Having a separate binary is useful as it allows sysadmins to remove
the buildkit functionality easily, if they are concerned about
increased attack surface / number of dependencies etc.

As part of this change, the locking behaviour and root dir of our
singularity-builtkitd have been changed to avoid conflicts.

Boltdb is used by buildkit to track build history, caching, and other
things. Boltdb only supports one process accessing a DB file at a
time and employs exclusive locking to ensure this.

Previously, our own singularity-buildkitd was using the default
buildkitd root, and creating a lock file with a random component. It
was possible to start multiple singularity-buildkitd instances that
could sometimesdeadlock due to Boltdb locking. To fix this, use our
own buildkitd root under `~/.singularity.d` so we don't compete with a
system buildkit. Use a lock file with a fixed name, and wait for
any `singularity-buildkitd` that's already running to complete before
starting another.

If users want to make use of a system-wide buildkitd cache etc. then
they should run a buildkitd separately and set `BUILDKIT_HOST` so that
Singularity will use it.

Fixes #2419