fuse: permit SIF/Squashfs mount via FUSE in all native mode flows
Broaden the `--sif-fuse` / `sif fuse = yes` functionality so that it
can be used to mount SIF/SquashFS container images with FUSE in all
native mode flows. Previously this was only supported in non-setuid
user namespace dependent flows.

Mounts and cleanup are performed as below:

* The SquashFS image is mounted, using `squashfuse[_ll]`, into a nested
temporary directory by the launcher routines in the initial
unprivileged CLI process.
* The container is started via the setuid or unpriv userns starter.
* The starter spawns two unprivileged processes in the initial host
namespaces:
  - POST_START_HOST
  - CLEANUP_HOST
* After successful startup of a container, the POST_START_HOST process
is signalled by the MASTER process. It then performs a lazy
unmount (via fusermount) of the FUSE filesystem in the host
namespaces, and removes the temporary directory. The FUSE filesystem
remains mounted and accessible in the container namespaces until the
container exits.
* When the container exits, or fails to start, the CLEANUP_HOST
process is signalled by the MASTER process. It checks for the
existence of the temporary directory, and if it is still present
unmounts the FUSE fs and removes the dir. This should only happen when
container startup fails before the POST_START_HOST is able to do the
same.

This approach is fairly robust against SIGKILL. As long as the lazy
unmount in the host namespace has fired, a SIGKILL of the container
process / runtime parent process will not leave orphan mounts, FUSE
processes, or namespaces.

If processes are SIGKILL-ed during container startup, then orphan
mounts / FUSE processes may be left over. There is no easy way around
this without moving FUSE mounts deeper into the runtime engine, which
then requires FUSE3 (which would preclude later support for extfs FUSE
mounts via fuse2fs).