fix: ensure host cleanup runs when parent exits
When a container doesn't exit cleanly, or fails to start correctly,
then the master process won't write into the cleanup socket to
initiate cleanup in the host namespaces.

Prior to this commit, this caused the host cleanup to exit with an EOF
error, leaving FUSE mounts in place in host namespaces.

The host cleanup process has SIGTERM set as its parent death signal,
so we can trap this to ensure that cleanup runs for an 'unclean' exit
where the master doesn't initate cleanup via the socket.