e2e: gate OCI-mode / buildkitd on kernel >=4.18
We require FUSE in userns for OCI-mode and our singularity-buildkitd
(which uses a fuse-overlayfs snapshotter on kernel <5.11). FUSE in
userns is supported on kernel >=4.18.

Add the kernel requirement to the OCI profile.

Modify the buildkit / Dockerfile build tests to use the OCI profile
instead of specifying the `--oci` flag directly.

Fixes #2494
Fixes #2493