fix: singularity-buildkitd: Don't use --keep for RUN steps
In our `singularity-buildkitd` executor code, we are incorporating the
pattern from the upstream `buildkitd`, where `RUN` steps are intended to
be executed with `runc` using the `--keep` flag. This leaves the
container running after the target process has completed. The container
must then be shutdown separately with a `runc delete` if it is not
inside of a userns.

Use of `--keep` and a separate delete was added to upstream `buildkitd`
to allow per-step resource tracking, this is not something we really
need for our provided ephemeral `singularity-buildkitd`.

Different to upstream `buildkitd`, our executor code may user `crun`
instead of `runc.` `crun` doesn't have the `--keep` flag, so logic
was added to avoid using it. However, there was a logic error in the
setting of the `isRunc` boolean leading to it being always false,
and `--keep` was never used even with `runc`.

On cgroups v1 systems, building as root, the lack of `--keep` results
in a subsequent `runc delete` call failing.

Given that we don't need per-step resource tracking, the safest solution
here is to always run without `--keep`, removing the need for a separate
delete step.

Fixes #2484