e2e: change OCSPResponder gating for SLES12
As well as EL7, SLES12 doesn't support some of the algorithms in
the OCSP chain that is tested.

Adjust the kernel requirement, so that it exclude SLES12 as well
as EL7.

Using a kernel requirement is an inexact / indirect method of
checking for algorithm support, but is simple and works across our
target supported distros.