fix: oci: always use rootless uid for inner uid map (release-4.1)
In OCI mode, as an unprivileged user, we have an outer UID/GID mapping
from the host UID/GID to 0:0 and an inner UID/GID mapping back to the host
UID.

The code was erroneously using the targetUID/GID for the container
process rather than the host UID/GID for the inner mapping. This meant
that when a container specifying an Dockerfile `USER` was run the mapping
and file owneship in the container was incorrect.

Fix the mapping and add e2e-tests that verify correct file ownership for
a container with a `USER`.

Fixes #2623