fix: oci: honor WORKDIR by default
The OCI image-spec `config.WorkingDir` must be used as the default
working directory for the container process. It is set by `WORKDIR` in a
Dockerfile build.

It's probably my fault, focusing too much on (incomplete) native mode
`--compat` behaviour rather than the OCI specs, that led to this omission.

Fixes #2663