fix: don't flatten layers with ggcr before extracting
The ggcr module has a bug in its `mutate.Extract` function, due to
optimisations that do not correctly track hard links that are
manipulated in specific ways across layers:

https://github.com/google/go-containerregistry/issues/977

Prior to this PR, we were using `mutate.Extract` to flatten an image so
that we were only extracting a single tar into a rootfs with umoci. This
meant that the bug above could cause umoci to fail, as the flattened tar
is invalid in certain circumstances.

Fix the issue by avoiding `mutate.Extract`. Instead extract each layer,
in order, using umoci.

Fixes #2684 (native mode)