fix: honor --writable for userns temp sandbox
Restore the long standing previous behaviour of singularity's native
runtime, where `--writable` can be used when running a SIF / SquashFS
container image in userns (`-u`) mode.

In this situation:

 - The image is extracted to a temporary sandbox.
 - The resulting container run from the sandbox is writable.

This behaviour has always been a bit confusing / inconsistent, because no
writes happen to the actual image that the user specified. Only the
temporary sandbox, which is deleted when the container exits, is written
to. The PR adds awarning to this effect.

We may want to revisit this behaviour at a future major version, because
it is inconsistent with other flows. However, this PR is needed because
not honoring `--writable` in this manner was a behaviour change regression
within a major version.

PR also adds additional comments around `launcher.prepareImage` to try
and clarify what's expected.

Fixes #2690