fix: store docker.io credential like index.docker.io
For historical reasons, the auth file key for docker hub is stored by
docker as:

  https://index.docker.io/v1/

We were successfully doing this for logins to `index.docker.io`, but
`docker.io` can also be used as an alias... and was not being mutated to
the above string.

We now have the relevant constants in our code, as they are not all
exported by ggcr - some are private there.

PR also adds a check that E2E_DOCKER_PASSWORD is not empty if
E2E_DOCKER_USERNAME is set. This tripped me up while chasing down the
underlying issue of E2E tests using auth on "docker.io" not working as
expected.

Fixes #2778