deps: switch to docker/docker/profile/seccomp
In order to drop container/common as a direct dependency, switch seccomp
profile loading & the default profile to use docker/docker code.

We are committed to docker/docker as a dependency given we are now using
buildkit etc. so we will always need to deal with breaking changes and
security issues in that ecosystem.

Dropping containers/common avoids us being tied to breaking changes /
security issues in the podman dependency ecosystem as well.